requirements.txt
# Discord Bot
discord.py>=2.3.2
python-dotenv>=1.0.0

# 音声処理
openai>=1.10.0
pydub>=0.25.1
ffmpeg-python>=0.2.0

# AI/ML関連
aiohttp>=3.9.1
numpy>=1.24.3
pandas>=2.0.3

# ユーティリティ
psutil>=5.9.5
python-dateutil>=2.8.2
pytz>=2023.3

# データベース（オプション）
aiosqlite>=0.19.0

# ログ・監視
colorlog>=6.8.0

# 開発用（オプション）
black>=23.12.0
flake8>=6.1.0
pytest>=7.4.4
pytest-asyncio>=0.21.1
